2.1.10 removed js warning and photoset layout limit

+++ public/includes/theme-helpers.php
@@ -362,7 +362,7 @@ function aesop_gallery_component_data_atts( $postid = '' ) {
 		'transition'  => sanitize_text_field( trim( $thumb_trans ) ),
 		'speed'   => $thumb_speed,
 		'thumbhide' => $thumb_hide,
-		'pslayout'  => (int) trim( $photoset_layout ),
+		'pslayout'  =>  trim( $photoset_layout ),
 		'pslightbox' => sanitize_text_field( trim( $photoset_lb ) ),
 		'content' => sanitize_text_field( trim( $hero_content ) ),
 	);
